relocate compose tests
